standardise project structure, add .gitignore and .gitattributes
