+++ .gitattributes
@@ -1,0 +1,2 @@
+# always normalise line endings
+* text=auto
