+++ StardewMods.sln
@@ -1,0 +1,32 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.27703.2035
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SauvignonInStardew", "SauvignonInStardew\SauvignonInStardew.csproj", "{9462B13A-0E4A-422D-8194-D8ECB8A27232}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "root", "root", "{8CD749FC-D224-4B31-8B8B-E00A20222691}"
+	ProjectSection(SolutionItems) = preProject
+		.gitattributes = .gitattributes
+		.gitignore = .gitignore
+		README.md = README.md
+	EndProjectSection
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{9462B13A-0E4A-422D-8194-D8ECB8A27232}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9462B13A-0E4A-422D-8194-D8ECB8A27232}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9462B13A-0E4A-422D-8194-D8ECB8A27232}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9462B13A-0E4A-422D-8194-D8ECB8A27232}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {888D4C1F-3643-4299-91EB-8A6EAE732BFE}
+	EndGlobalSection
+EndGlobal
