--- manifest.json
@@ -1,10 +1,0 @@
-{
-  "Name": "Sauvignon in Stardew",
-  "Author": "Jesse",
-  "Version": "1.0.0",
-  "Description": "Uses SMAPI. Adds a new building to the Carpenter Menu.",
-  "UniqueID": "Jesse.SauvignonInStardew",
-  "EntryDll": "Sauvignon in Stardew.dll",
-  "MinimumApiVersion": "2.0",
-  "UpdateKeys": ["Nexus:2597"]
-}
